Update priority in EPP flow control from uint to int (#1518)
* Update type for priority from uint to int in EPP flow control

* Update tests to accomodate for priority changes

* Avoid using Reverse to sort in descending order, update comments

+++ pkg/epp/flowcontrol/controller/internal/filter_test.go
@@ -128,7 +128,7 @@ func TestSubsetPriorityBandAccessor(t *testing.T) {
 
 	t.Run("should pass through priority and name", func(t *testing.T) {
 		t.Parallel()
-		assert.Equal(t, uint(10), subsetAccessor.Priority(), "Priority() should pass through from the original accessor")
+		assert.Equal(t, 10, subsetAccessor.Priority(), "Priority() should pass through from the original accessor")
 		assert.Equal(t, "High", subsetAccessor.PriorityName(),
 			"PriorityName() should pass through from the original accessor")
 	})
